Implement route grouping for organizing routes
