+++ app/(dashboard)/about/page.jsx
@@ -1,6 +1,6 @@
-import "../../styles/About.css";
+import "../../../styles/About.css";
 
-export default function Home() {
+export default function About() {
   return (
     <main className="about">
       <h1>Your Story Starts With Us.</h1>
