+++ app/(dashboard)/blog/[id]/page.jsx
@@ -2,8 +2,8 @@
 
 import { useState, useEffect } from "react";
 import Image from "next/image";
-import blogImage from "../../../public/assets/blog.jpg";
-import "../../../styles/Post.css";
+import blogImage from "../../../../public/assets/blog.jpg";
+import "../../../../styles/Post.css";
 
 const URL = "https://dummyjson.com/posts";
