+++ app/(dashboard)/layout.jsx
@@ -1,17 +1,15 @@
-import "./globals.css";
+import Header from "../../components/header/Header";
+import Footer from "../../components/footer/Footer";
 
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
+export default function DashboardLayout({ children }) {
   return (
     <html lang="en">
       <body>
         <div className="App">
           <div className="container">
+            <Header />
             <div className="pages">{children}</div>
+            <Footer />
           </div>
         </div>
       </body>
