+++ app/login/page.jsx
@@ -1,54 +1,16 @@
 "use client";
 
-import { useState } from "react";
-import { useRouter } from "next/navigation";
+
+import LoginForm from "../../components/LoginForm/LoginForm";
 import "../../styles/Login.css";
 
-function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const router = useRouter();
+export default function Login() {
 
-  const handleLogin = async () => {
-    console.log("Login attempt with:", email, password);
-    router.push("/");
-  };
 
   return (
     <div className="login-container">
       <h1>Login</h1>
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          handleLogin();
-        }}
-      >
-        <div className="input-group">
-          <label htmlFor="email">Email</label>
-          <input
-            type="email"
-            id="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </div>
-        <div className="input-group">
-          <label htmlFor="password">Password</label>
-          <input
-            type="password"
-            id="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </div>
-        <button type="submit" className="button">
-          Login
-        </button>
-      </form>
+      <LoginForm />
     </div>
   );
 }
-
-export default Login;
