+++ styles/page.css
@@ -15,6 +15,7 @@
   border: 1px solid #ccc;
   padding: 20px;
   text-align: center;
+  cursor: pointer;
 }
 
 .product img {
